feat(fuzz): add DNS upstream-response parsing fuzz target
Adds a `dns_response` ClusterFuzzLite target covering the upstream
DNS-response trust boundary now that full forwarding/resolution is in
place. Every datagram the forwarders receive from an upstream resolver
(or an off-path attacker who beats it back to our socket) is
wire-parsed with `Message::from_bytes` and walked by our caching
classifier before it can be relayed or cached; the target fuzzes that
parse+classify core directly, with no socket or runtime.

The classification core (`classify_response`) moves down from
`wardnetd/dns/pipeline.rs` into a new `wardnetd-services/dns/response`
module, next to `DnsCache` — the type its TTL output feeds. It is a
pure function over hickory-proto types, so the move is behaviour-
preserving and improves cohesion. Crucially it keeps the fuzz target
off the `wardnetd` binary crate, which drags in `rustables`
(nftables/bindgen) and won't build in the OSS-Fuzz image; the target
now links only the already-fuzz-safe `wardnetd-services`.

- wardnetd-services: new `dns::response` module with `classify_response`
  and, behind the existing `fuzz-internals` feature, the synchronous
  `parse_and_classify_response_for_fuzz` seam. pipeline.rs delegates to it.
- fuzz workspace: new `dns_response` bin over wardnetd-services.
- build.sh: register `dns_response` in FUZZ_TARGETS.
- README + scheduled workflow: document the new target and drop the
  "DNS response parsing — once DNS forwarding is implemented" candidate
  now that it is implemented.
