+++ source/daemon/crates/wardnetd-services/src/dns/mod.rs
@@ -6,6 +6,7 @@ pub mod dhcp_lan_runner;
 pub mod filter_parser;
 pub mod log_sink;
 pub mod query_log_runner;
+pub mod response;
 pub mod runner;
 pub mod server;
 pub mod service;
@@ -16,6 +17,7 @@ pub use capture_runner::DnsCaptureRunner;
 pub use dhcp_lan_runner::DhcpLanRunner;
 pub use log_sink::{DnsLogSink, DnsLogSinkChannels, row_to_event};
 pub use query_log_runner::DnsQueryLogRunner;
+pub use response::classify_response;
 pub use runner::DnsRunner;
 pub use server::{DnsServer, DnsSocket};
 pub use service::{DnsService, DnsServiceImpl};
